Declaring NaturalSort as a NuGet dependency.
Since NaturalSort is on NuGet, let's import the project as a NuGet
dependency rather than a project reference.

This project reference will likely live in a separate repository at
some point so that a user interested in the natural sort algorithm
wouldn't have to download the IQueryable project as well.